pyproject.toml
[tool.poetry]
name = "agentartificial"
version = "0.1.0"
description = "Agent Artificial's OpenAI compatible api client"
authors = ["Richard <richard@agentartificial.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.14.2"
python-dotenv = "^1.0.1"
websocket = "^0.2.1"
httpx = "^0.27.0"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
